Extracted common functionality of WALA/OPAL plugins to AnalyzerPlugin
