+++ analyzer/javacg-wala/src/test/java/eu/fasten/analyzer/javacgwala/data/callgraph/WalaResultAnalyzerTest.java
@@ -28,8 +28,10 @@ import eu.fasten.core.data.FastenJavaURI;
 import java.io.File;
 import java.io.IOException;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
+@Disabled("Disabled until a way to make Wala platform independent found")
 class WalaResultAnalyzerTest {
 
     private static CallGraph graph;
